Updated backend doc: Covers steps to deploy/connect MySQL and SpringBoot Docker Containers
- Also makes sure to have user run database schema
creation script on their host machine from mysql
cli, then create a named volume outside the scope
of any container and then launch mysql container
that the springboot container will later be able
to connect to since when we launch the springboot
container, we include env variable arguments
to connect to mysql container using JDBC MySQL
URL and username and password. Just need to do
further testing on this connection between
MySQL and SpringBoot Docker Containers.

- Also SpringBoot container connects to
a BackEnd docker network and a FrontEnd
docker network. The BackEnd network is a bridged
network for SpringBoot to communicate with
MySQL container. The FrontEnd network is a bridged
network for SpringBoot to communicate with
ReactJS container.