Added data insertion for ct_grant data and resolved many syntax issues
- Able to insert data into ct_grant table. Notice the trial_ids
match with the trial_ids in the trial table. I think Mu incremented
trial_id to make it easier to follow compared to the previous
version of trial_ids from original clinical trials excel
sheet. The trial_ids are easier to follow with her approach.
- You'll notice some trial_ids in the ct_grant table are
NULL indicating that they do not exist yet in the trial
table. There can be grant money out there that doesnt have an
associating trial_id yet.
- Fixed syntax issues coming from special characters that needed
escape sequences such as \', ", etc.
- Added 1000+ records for insertion to ct_grant table.
- Updated ct_grant table's creation for almost each columns amount
of characters that can be in the value based on the boundary limits
that I saw when I was inserting data. Now that boundary limit
shouldn't be an issue with our sample data.
- start_data and end_date have been changed from date to varchar
since the value is a string to represent the date and not
year month day.
- also ct_grant's foreign key trial_id references trial table's
primary key trial_id, that foreign key on delete is set to default
value NULL.
- Also referenced Somil's Pull Request and saw some updates on
his updated version of create table that I included in this approach.
- But majority of solving this insertion data issue was through
trial and error and following mysql error log.